+++ kuma-client/src/models/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod event;
 pub(crate) mod response;
 
+pub mod docker_host;
 pub mod maintenance;
 pub mod monitor;
 pub mod notification;